// ADL/bbog-bv-ionic/src/app/models/column-options-dynamic.ts
export interface ColumnOptionsDynamic {
    header: string;
    class?: any;
    cell?: any;
    position?: string;
    lg?: string;
    xl?: string;
    btn1?: boolean;
    img?: boolean;
    tablet?: boolean;
    mobile?: boolean;
}

export interface ExpandableData {
    title: string;
    cell: any;
    class?: any;
    hiddenWeb?: boolean;
}